Fix mappers and dto's

+++ src/main/java/com/matchvagas/backend/dto/TipoNotificacaoResponseDTO.java
@@ -4,6 +4,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record TipoNotificacaoResponseDTO(
-    Integer id,
-    String status
+        Long id,        // era: Integer — corrigido para Long (padrão do projeto)
+        String status
 ) {}
